add logs to retry
