+++ driver_test.go
@@ -2,6 +2,8 @@
 
 package gosnowflake
 
+//lint:file-ignore U1000 Ignore all unused code
+
 import (
 	"bytes"
 	"context"
